periodogram roadblock
Made all cosmetic and textual changes. Unfortunately when I re-ran the cells, it seems that lightkurve's periodogram function is broken. I have tried to run it outside of this lesson locally both in notebooks and IDEs but it won't seem to run anywhere. The error is: "AttributeError: module 'astropy.timeseries.periodograms.lombscargle_multiband.implementations.main' has no attribute '_is_regular'". I need someone else to confirm if they are having this issue and to please advise.